requirements.txt
fastapi
uvicorn
loguru
ncatbot==4.0.7
py-jsonl
pyecharts